% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{on_municipal_boundaries}
\alias{on_municipal_boundaries}
\alias{on_municipal_boundaries_dry}
\alias{on_municipal_boundaries_wet}
\title{Ontario Municipal Boundaries}
\format{
\subsection{\code{on_municipal_boundaries}}{

A 'simple features collection' (sf) with 657 rows and 4 columns:
\describe{
\item{\code{munid} \code{<chr>}}{Municipality ID (linkable)}
\item{\code{municipality} \code{<chr>}}{Municipality name (esoteric)}
\item{\code{tier_code} \code{<fct>}}{
Tier code for the municipality. One of:
\describe{
\item{\code{D}}{District}
\item{\code{UT}}{Upper Tier Municipality}
\item{\code{LT}}{Lower Tier Municipality}
\item{\code{ST}}{Single Tier Municipality}}}
\item{\code{geometry} \code{<sfc_MULTIPOLYGON>}}{Geometry of the municipality}}
}

An object of class \code{sf} (inherits from \code{grouped_df}, \code{tbl_df}, \code{tbl}, \code{data.frame}) with 444 rows and 4 columns.

An object of class \code{sf} (inherits from \code{grouped_df}, \code{tbl_df}, \code{tbl}, \code{data.frame}) with 445 rows and 4 columns.
}
\source{
Land Information Ontario (LIO) via Ontario GeoHub:
\describe{
\item{LT/ST}{
\href{https://geohub.lio.gov.on.ca/datasets/64fb702e16204c3e88b528d9759f1174_14/about}{Municipal Boundaries (Lower and Single Tier)}
}
\item{UT/D}{
\href{https://geohub.lio.gov.on.ca/datasets/11be9127e6ae43c4850793a3a2ee943c_13/about}{Municipal Boundaries (Upper and District)}
}
}
}
\usage{
on_municipal_boundaries_dry

on_municipal_boundaries_wet
}
\description{
Raw data, available from the source links below, represent boundaries at a much higher resolution,
and include separate entries for mainland, island, and water areas within municipalities.
The combined data are simplified using \code{rmapshaper::ms_simplify} with a \code{keep} value of 0.001.
After simplification, boundaries internal to municipalities are dissolved using \code{rmapshaper::ms_dissolve}.

code{on_municipal_boundaries_dry} land-only boundaries for all Ontario municipalities

\code{on_municipal_boundaries_wet} combined land and water boundaries for all Ontario municipalities
}
\details{
Simplified boundaries for all Ontario municipalities.
}
\keyword{datasets}
