% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onbound-package.R
\docType{package}
\name{onbound-package}
\alias{onbound}
\alias{onbound-package}
\title{onbound: Ontario Municipal Boundary Data}
\description{
Provides convenient access to Upper, Lower, and Single Tier municipal boundary data for Ontario, Canada. Data are sourced from the Ontario Ministry of Municipal Affairs and Housing, by way of Ontario Open Data. They are provided in the form of sf data frames, and are intended to be used in conjunction with the `sf` package.
}
\author{
\strong{Maintainer}: Thomas McManus \email{tmcmanus@blueprint-ade.ca}

}
\keyword{internal}
